Added chatbox functionality and initial OpenAI API call
Current issue that needs working on: right now the API call takes place on the browser side which is no bueno. We need to create an endpoint on the server side so we can securely send the API key back and forth. I did test this with a temp key though and it works! Yayyyy